- Adding trade list feature almost completed.

+++ app/src/main/java/com/jeepchief/clubhouse/model/database/traderecord/TradeRecordEntity.kt
@@ -19,5 +19,5 @@ data class TradeRecordEntity(
     var grade: Int,
 
     @ColumnInfo
-    var value: Int
+    var value: Long
 )
